MD to PDF
I wanted a way to bookmark where I was in the document.  I did not find
any method of embedding a bookmark in .md files, so I converted all .md
files to .pdf to allow for bookmarking.  (I used pandoc to do this and
looked at jandoc to do a batch job, but it does not support recursive
searches)...